+++ megaavr/tools/libs/pymcuprog/serialupdi/link.py
@@ -44,14 +44,20 @@ class UpdiDatalink:
             if not self._check_datalink():
                 raise PymcuprogError("UPDI initialisation failed")
 
-    def change_baud(self, baud):
+    def change_baud(self, baud, thirtytwoisoption=False):
         if self.updi_phy is not None:
             self.stcs(constants.UPDI_CS_CTRLA, 0x06)
             if baud <= 115200:
+                self.logger.info("Setting UPDI clock to 4 MHz")
                 self.stcs(constants.UPDI_ASI_CTRLA, 0x03)
-            elif baud > 230400:
+            elif thirtytwoisoption and baud > 921600:
+                self.logger.info("Setting UPDI clock to 32 MHz - Good luck!")
+                self.stcs(constants.UPDI_ASI_CTRLA, 0x01)
+            elif baud >= 460800:
+                self.logger.info("Setting UPDI clock to 16 MHz")
                 self.stcs(constants.UPDI_ASI_CTRLA, 0x01)
             else:
+                self.logger.info("Setting UPDI clock to 8 MHz")
                 self.stcs(constants.UPDI_ASI_CTRLA, 0x02)
             self.updi_phy.change_baud(baud)
 
@@ -64,7 +70,7 @@ class UpdiDatalink:
                 self.logger.info("UPDI init OK")
                 return True
         except PymcuprogError:
-            self.logger.warning("Check failed")
+            self.logger.warning("UPDI init failed: Can't read CS register.")
             return False
         self.logger.info("UPDI not OK - reinitialisation required")
         return False
@@ -74,7 +80,7 @@ class UpdiDatalink:
         Load data from Control/Status space
         :param address: address to load
         """
-        self.logger.debug("LDCS from 0x%02X", address)
+        self.logger.info("LDCS from 0x%02X", address)
         self.updi_phy.send([constants.UPDI_PHY_SYNC, constants.UPDI_LDCS | (address & 0x0F)])
         response = self.updi_phy.receive(self.LDCS_RESPONSE_BYTES)
         numbytes_received = len(response)
@@ -90,7 +96,7 @@ class UpdiDatalink:
         :param address: address to store to
         :param value: value to write
         """
-        self.logger.debug("STCS to 0x%02X", address)
+        self.logger.info("STCS %02X to 0x%02X", value, address)
         self.updi_phy.send([constants.UPDI_PHY_SYNC, constants.UPDI_STCS | (address & 0x0F), value])
 
     def ld_ptr_inc(self, size):
@@ -129,6 +135,10 @@ class UpdiDatalink:
         response = self.updi_phy.receive(1)
 
         if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
+            if len(response > 0):
+                self.logger.error("Expecting ACK after ST8 *ptr++. Got %d.", num, request[0])
+            else:
+                self.logger.error("Expecting ACK after ST8 *ptr++. Got nothing.", num)
             raise PymcuprogError("ACK error with st_ptr_inc")
 
         num = 1
@@ -137,6 +147,10 @@ class UpdiDatalink:
             response = self.updi_phy.receive(1)
 
             if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
+                if len(response > 0):
+                    self.logger.error("Expecting ACK after ST8 *ptr++, after byte %d. Got 0x%00X}", num, request[0])
+                else:
+                    self.logger.error("Expecting ACK after ST8 *ptr++, after byte %d. Got nothing", num)
                 raise PymcuprogError("Error with st_ptr_inc")
             num += 1
 
@@ -151,7 +165,11 @@ class UpdiDatalink:
         response = self.updi_phy.receive(1)
 
         if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
-            raise PymcuprogError("ACK error with st_ptr_inc16")
+            if len(response > 0):
+                self.logger.error("Expecting ACK after ST16 *ptr++. Got {}}", response[0])
+            else:
+                self.logger.error("Expecting ACK after ST16 *ptr++. Got nothing")
+            raise PymcuprogError("Error with st_ptr_inc16")
 
         num = 2
         while num < len(data):
@@ -159,6 +177,10 @@ class UpdiDatalink:
             response = self.updi_phy.receive(1)
 
             if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
+                if len(response > 0):
+                    self.logger.error("Expecting ACK after ST16 *ptr++, after word %d.  0x%00X}", num, request[0])
+                else:
+                    self.logger.error("Expecting ACK after ST16 *ptr++, after word %d. Got nothing", num)
                 raise PymcuprogError("Error with st_ptr_inc16")
             num += 2
 
@@ -166,7 +188,7 @@ class UpdiDatalink:
         """
         Store a 16-bit word value to the pointer location with pointer post-increment
         :param data: data to store
-        :blocksize: max number of bytes being sent -1 for all.
+        :blocksize: max number of bytes being sent, None for all.
                     Warning: This does not strictly honor blocksize for values < 6
                     We always glob together the STCS(RSD) and REP commands.
                     But this should pose no problems for compatibility, because your serial adapter can't deal with 6b chunks,
@@ -178,7 +200,7 @@ class UpdiDatalink:
         repnumber= ((len(data) >> 1) -1)
         data = [*data, *[constants.UPDI_PHY_SYNC, constants.UPDI_STCS | constants.UPDI_CS_CTRLA, 0x06]]
 
-        if blocksize == -1 :
+        if blocksize is None:
             # Send whole thing at once stcs + repeat + st + (data + stcs)
             blocksize = 3 + 3 + 2 + len(data)
         num = 0
@@ -195,11 +217,45 @@ class UpdiDatalink:
                             *[constants.UPDI_PHY_SYNC, constants.UPDI_ST | constants.UPDI_PTR_INC | constants.UPDI_DATA_16],
                             *data[:blocksize - 8]]
             num = blocksize - 8
+            if len(firstpacket) == 64 and blocksize != 64:
+                firstpacket = firstpacket[:32]
+                num = 32-8
+                # workaround bug in D11C as serial adapter compiled with the USB implementation used in the mattairtech core;
+                # this chokes on any block of exactly (!!) 64 bytes. Nobody seems to understand the reason for this bizzare issue.
+                # The D11C as serial adapter is important because the fablab at MIT uses them heavilly, and wishes to upload Arduino sketches to tinyAVR 0/1/2-series
+                # and AVR-Dx-series parts with them. It was desirable from their perspective to have an more accessible firmware (like that Arduino one) as opposed
+                # to a pure-C one. Hence, a workaround is implemented to avoid triggering it. Thankfully, 64-byte blocks being written are not as natural as one might
+                # expect for a power-of-two, since we combine the data with the commands on either side.
+                # This workaround will only be invoked under unusual conditions, specifically if one of these is true:
+                #   A. The last page of the hex file has a length of 53 bytes exactly OR
+                #   B. The last page of the hex file has a length modulo the write chunk size, of 53 bytes exactly OR
+                #   C. A write chunk is specified which is exactly 53 bytes shorter than the page OR
+                #   D. The write chunk is specified as 153 when writing to a part with a page size of 512 OR
+                #   E. Situations like D in the event that a future product permits REPEAT with 2 bytes of data to support pages larger than 512b.
+                # Conditions C-E will invoke it on the last chunk of every page; For E, here are 6 cursed numbers for 1024b pages with 2 byte repeats, and 2 for 2048b pages
+                # (in addition to pagelength - 52), however, there is no specific reason to expect such a part will be made. In any event, for those conditions, the
+                # performance impact is on the order of 1-4ms per page, the same as the the penalty for each chunk that a write is divided into, hence the speed penalty of
+                # using that write chunk size will be (1-4ms * ceil(page size/chunk size) + 1 instead of 1-4ms * ceil(page size/chunk size), which is always smaller than the
+                # penalty already being accepted for the write chunking.
+                # Conditions A and B will happen at most once per upload; thus the performance penalty would be acceptable even if they happened frequently. However
+                # avr-gcc rarely, if ever, generates odd size hex files. I was not able to get it to by tweaking the sketch I was compiling, so they will occur on
+                # considerably less than 1/64th of the time, if it is even possible for avr-gcc to generate such a file, which is unclear.
+                # In summary, the performance impact is small but noticable in remote corner cases (where chunk size is chosen randomly, or intentionally in an effort to provoke
+                # this bug, and is negligible (at worst 4 ms for 1 out of 32 uploads, assuming upload size modulo page size are evenly distributed and even ) in unlikely cases
+                # (requiring an odd write chunk size such that B could be provoked from an even-length hex file) as well as -possibly- case A if there is indeed a programming
+                # construct that results in a non-even length hex file. It is (barely) worth noting that for a hypothetical part that used a 2 byte REPEAT argument, this could
+                # be provoked with an even sketch size, and so would impose a penalty of less than 4ms on 1 out of 32 uploads.
+                # In light of the fact write chunking is intended as a workaround for an ill-mannered serial adapter in the first place (ex: HT42 workaround in DxCore), this
+                # performance impact is not a concern.
+                # If a write chunk size of 64 is specified, we will not activate this workaround.
+                # -Spence 6/29/21
         self.updi_phy.send( firstpacket )
 
         # if finite block size, this is used.
         while num < len(data):
             data_slice = data[num:num+blocksize]
+            if len(data_slice) == 64 and blocksize != 64:
+                data_slice=data[num:num+32] # workaround as above.
             self.updi_phy.send(data_slice)
             num += len(data_slice)
 
@@ -243,11 +299,19 @@ class UpdiDatalink:
         """
         response = self.updi_phy.receive(1)
         if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
+            if len(response >= 0):
+                self.logger.error("expecting ACK after ST, but got: %02x", response[0])
+            else:
+                self.logger.error("expecting ACK after ST, got nothing.")
             raise PymcuprogError("Error with st")
 
         self.updi_phy.send(values)
         response = self.updi_phy.receive(1)
         if len(response) != 1 or response[0] != constants.UPDI_PHY_ACK:
+            if len(response >= 0):
+                self.logger.error("expecting ACK after ST value, but got: %02x", response[0])
+            else:
+                self.logger.error("expecting ACK after ST value, got nothing.")
             raise PymcuprogError("Error with st")
 
 
@@ -268,7 +332,7 @@ class UpdiDatalink16bit(UpdiDatalink):
         :param address: address to load from
         :return: value read
         """
-        self.logger.info("LD from 0x{0:06X}".format(address))
+        self.logger.debug("LD from 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_LDS | constants.UPDI_ADDRESS_16 | constants.UPDI_DATA_8,
              address & 0xFF, (address >> 8) & 0xFF])
@@ -280,7 +344,7 @@ class UpdiDatalink16bit(UpdiDatalink):
         :param address: address to load from
         :return: values read
         """
-        self.logger.info("LD from 0x{0:06X}".format(address))
+        self.logger.debug("LD from 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_LDS | constants.UPDI_ADDRESS_16 | constants.UPDI_DATA_16,
              address & 0xFF, (address >> 8) & 0xFF])
@@ -293,7 +357,7 @@ class UpdiDatalink16bit(UpdiDatalink):
         :param address: address to write to
         :param value: value to write
         """
-        self.logger.info("ST to 0x{0:06X}".format(address))
+        self.logger.debug("ST to 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_STS | constants.UPDI_ADDRESS_16 | constants.UPDI_DATA_8,
              address & 0xFF, (address >> 8) & 0xFF])
@@ -305,7 +369,7 @@ class UpdiDatalink16bit(UpdiDatalink):
         :param address: address to write to
         :param value: value to write
         """
-        self.logger.info("ST to 0x{0:06X}".format(address))
+        self.logger.debug("ST to 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_STS | constants.UPDI_ADDRESS_16 | constants.UPDI_DATA_16,
              address & 0xFF, (address >> 8) & 0xFF])
@@ -316,7 +380,7 @@ class UpdiDatalink16bit(UpdiDatalink):
         Set the pointer location
         :param address: address to write
         """
-        self.logger.info("ST to ptr")
+        self.logger.debug("ST to ptr")
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_ST | constants.UPDI_PTR_ADDRESS | constants.UPDI_DATA_16,
              address & 0xFF, (address >> 8) & 0xFF])
@@ -342,7 +406,7 @@ class UpdiDatalink24bit(UpdiDatalink):
         :param address: address to load from
         :return: value read
         """
-        self.logger.info("LD from 0x{0:06X}".format(address))
+        self.logger.debug("LD from 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_LDS | constants.UPDI_ADDRESS_24 | constants.UPDI_DATA_8,
              address & 0xFF, (address >> 8) & 0xFF, (address >> 16) & 0xFF])
@@ -354,7 +418,7 @@ class UpdiDatalink24bit(UpdiDatalink):
         :param address: address to load from
         :return: values read
         """
-        self.logger.info("LD from 0x{0:06X}".format(address))
+        self.logger.debug("LD from 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_LDS | constants.UPDI_ADDRESS_24 | constants.UPDI_DATA_16,
              address & 0xFF, (address >> 8) & 0xFF, (address >> 16) & 0xFF])
@@ -367,7 +431,7 @@ class UpdiDatalink24bit(UpdiDatalink):
         :param address: address to write to
         :param value: value to write
         """
-        self.logger.info("ST to 0x{0:06X}".format(address))
+        self.logger.debug("ST to 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_STS | constants.UPDI_ADDRESS_24 | constants.UPDI_DATA_8,
              address & 0xFF, (address >> 8) & 0xFF, (address >> 16) & 0xFF])
@@ -379,7 +443,7 @@ class UpdiDatalink24bit(UpdiDatalink):
         :param address: address to write to
         :param value: value to write
         """
-        self.logger.info("ST to 0x{0:06X}".format(address))
+        self.logger.debug("ST to 0x{0:06X}".format(address))
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_STS | constants.UPDI_ADDRESS_24 | constants.UPDI_DATA_16,
              address & 0xFF, (address >> 8) & 0xFF, (address >> 16) & 0xFF])
@@ -390,7 +454,7 @@ class UpdiDatalink24bit(UpdiDatalink):
         Set the pointer location
         :param address: address to write
         """
-        self.logger.info("ST to ptr")
+        self.logger.debug("ST to ptr")
         self.updi_phy.send(
             [constants.UPDI_PHY_SYNC, constants.UPDI_ST | constants.UPDI_PTR_ADDRESS | constants.UPDI_DATA_24,
              address & 0xFF, (address >> 8) & 0xFF, (address >> 16) & 0xFF])