+++ megaavr/tools/libs/pymcuprog/utils.py
@@ -155,7 +155,7 @@ def compare(data0, data1, offset, verify_mask=None):
 
     # Check first that lengths match
     if len(data0) != len(data1):
-        raise ValueError("Length mismatch on verify")
+        raise ValueError("Length mismatch on verify, expect 0x{:04X} but got 0x{:04X}".format(len(data0),len(data1)))
 
     mask_len = len(verify_mask)
 
@@ -255,7 +255,7 @@ def enum(**enums):
     return type('Enum', (), enums)
 
 
-def verify_flash_from_bin(bin_filename, backend, offset=0):
+def verify_flash_from_bin(bin_filename, backend, offset=0, max_read_chunk=None):
     """
     Verify the contents of flash against a bin-file
 
@@ -269,7 +269,7 @@ def verify_flash_from_bin(bin_filename, backend, offset=0):
     for line in bin_file.readlines():
         bin_data.extend(line)
 
-    verify_status = backend.verify_memory(bin_data, 'flash', offset)
+    verify_status = backend.verify_memory(bin_data, 'flash', offset, max_read_chunk=max_read_chunk)
     if verify_status is False:
         return False
     return True